refactor move dashboard snapshot storage to shared
